Massively restructure codebase; more refactors incoming.
